# code-with-quarkus/src/main/java/at/freelenzer/invoice/model/Tax.java
package at.freelenzer.invoice.model;

import jakarta.persistence.Embeddable;

@Embeddable
public class Tax {
    // Tax rate in percent
    public float taxRate;
    // Total from what tax is calculated
    public float taxTotal;
    // Amount of tax
    public float taxAmount;
}